% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tazuku.R
\name{generateTakuzuGrid}
\alias{generateTakuzuGrid}
\title{Generating Takuzu Grid}
\usage{
generateTakuzuGrid(size)
}
\arguments{
\item{size}{Grid size (eg 8 for 8x8)}
}
\value{
Generated Takuzu Mesh Matrix
}
\description{
Generating Takuzu Grid
}
