% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tazuku.R
\name{get_difficulty_params}
\alias{get_difficulty_params}
\title{Get Difficulty Level Parameters}
\usage{
get_difficulty_params(difficulty)
}
\arguments{
\item{difficulty}{A string indicating the difficulty level ("easy", "medium", "hard", "expert").}
}
\value{
A list containing the fill percentage and chaotic flag.
}
\description{
This function retrieves the parameters for a given difficulty level.
}
