% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tazuku.R
\name{generateTakuzuGrid}
\alias{generateTakuzuGrid}
\title{Generating Takuzu Grid}
\usage{
generateTakuzuGrid(size = 8, difficulty = "easy")
}
\arguments{
\item{size}{Grid size (e.g., 6 for 6x6). Default is 8x8.}

\item{difficulty}{Level of difficulty ("easy", "medium", "hard", "expert"). Default is "easy".}
}
\value{
Generated Takuzu grid as a matrix
}
\description{
Generating Takuzu Grid
}
